requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-dotenv>=1.0.1
google-genai>=0.6.0
pinecone>=5.0.0
pinecone[grpc]>=5.0.0
requests>=2.31.0
ipykernel>=6.29.0
